Problem of Q and R
In the previous code, Q and R were not modified in the function main. This problem was circumvented by creating two function QR, one function QR, called "QR_givens_Q" gives the matrix Q and the other function QR, called, "QR_givens_R", determines R. However, the code still doesn't work Q and R have values which are too high and Q*R=!A. I think it may come from the function "mat_rot" because it is in radian and not in degrees.